+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-## 0.1.1 (Unreleased)
+## 0.2.0 (Unreleased)
+
+IMPROVEMENTS:
+ * New `scale-in` and `scale-out` commands  allow an operator to manually scale jobs and task groups based on counts or percentages [GH-172](https://github.com/jrasell/levant/pull/172)
+ * New template functions allowing the lookup of variables from Consul KVs as well as ISO-8601 timestamp generation [GH-175](https://github.com/jrasell/levant/pull/175) 
+ * Multiple variable files can be passed on each run, allowing for common configuration to be shared across jobs [GH-180](https://github.com/jrasell/levant/pull/180)
+ * Provide better command help for deploy and render commands [GH-183](https://github.com/jrasell/levant/pull/184)
+
+BUG FIXES:
+ * Fix formatting with version summary output which had erronous quote [GH-170](https://github.com/jrasell/levant/pull/170)
+
+## 0.1.1 (13 May 2018)
 
 IMPROVEMENTS:
  * Use govvv for builds and to supply additional version information in the version command output [GH-151](https://github.com/jrasell/levant/pull/151)